Separate entry point for node
